update check code
