+++ packages/shared/lib/components/PageHeader/style.ts
@@ -19,6 +19,18 @@ export const PageHeaderStyleWrapper = styled('div')`
       theme.sharedTheme.components.pageHeader.title.color};
   }
 
+  .refresh-icon {
+    width: 28px;
+    height: 28px;
+    display: flex;
+    justify-content: center;
+
+    svg {
+      color: ${({ theme }) => theme.sharedTheme.uiToken.colorTextTertiary};
+      font-size: 16px;
+    }
+  }
+
   &.fixed-style {
     position: fixed;
     width: calc(100% - 220px);
